Add placeholder for a sequential chapter number
Chapter numbers can often be fractional, for example 7.5. When saving such chapters, the alphabetical order is violated and
such chapters go earlier than the main ones (the file with chapter 7.5 before chapter 7). This patch adds the ability to add a
sequential chapter number to the file name to preserve order.
For example, template like a that:
--filename-chapter="{chapter.num} {chapter.name}{file_ext}"